final/router/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.27.3